requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
filelock==3.9.0
idna==3.3
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.13.1
jedi==0.18.2
joblib==1.2.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
pipenv==2023.2.18
platformdirs==3.0.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.2
requests==2.26.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.7
virtualenv==20.19.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
zipp==3.15.0
